synthetic_requirements.txt
openai>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
geopy>=2.3.0
python-dotenv>=1.0.0
requests>=2.28.0
